+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 - OpenAI: Backfill required 'query' field when service only provides 'queries'.
+- Agent Bridge: Add `model_event_sink` option for bridge to control order of model event emission (used to nest model events properly in agent spans).
 
 ## 0.3.225 (23 May 2026)